Started creating functions to interact with sqlite db
