+++ App.js
@@ -15,7 +15,8 @@ import WordList from './components/WordList';
 import { UserProvider, UserContext } from './context/UserContext';
 import { synchronizeUnlockedWordsForUser } from './firebase/getUnlockedWords';
 import colors from './styles/colors';
-import initDb from "./sqlite/sqlite";
+import initDb from "./sqlite/init-sqlite";
+import unlockNextWordForGuest from "./sqlite/unlockNextWordForGuest";
 
 const MainApp = () => {
     const [selectedComponent, setSelectedComponent] = useState('Home');
@@ -87,6 +88,7 @@ const MainApp = () => {
 
 const App = () => {
     initDb();
+    unlockNextWordForGuest('noun-das-mal');
 
     return (
         <UserProvider>
