+++ sqlite/updateWordScoreForGuest.js
@@ -1,0 +1,12 @@
+import * as SQLite from 'expo-sqlite';
+
+const updateWordScoreForGuest = async (wordId, newScore) => {
+    try {
+        const db = await SQLite.openDatabaseAsync('wordsDatabase.sqlite');
+        await db.runAsync(`UPDATE words SET score = ? WHERE id = ?`, newScore, wordId);
+    } catch (error) {
+        console.error(error);
+    }
+}
+
+export default updateWordScoreForGuest;
